+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## Unreleased
 
 - enh(android): Show link to issues page after setup is complete (#448)
-
+- fix(nextjs): Add selfhosted url in `next.config.js` (#438)
+  
 ## 3.12.0
 
 - feat(sourcemaps): Automatically insert Sentry Webpack plugin (#432)