+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - enh(android): Show link to issues page after setup is complete (#448)
+- fix(nextjs): Add selfhosted url in `next.config.js` (#438)
 - fix(nextjs): Create necessary directories in app router (#439)
 
 Work in this release contributed by @andreysam. Thank you for your contributions!