+++ CHANGELOG.md
@@ -4,7 +4,10 @@
 
 - enh(android): Show link to issues page after setup is complete (#448)
 - fix(nextjs): Add selfhosted url in `next.config.js` (#438)
-  
+- fix(nextjs): Create necessary directories in app router (#439)
+
+Work in this release contributed by @andreysam. Thank you for your contributions!
+
 ## 3.12.0
 
 - feat(sourcemaps): Automatically insert Sentry Webpack plugin (#432)